third_party/blink/web_tests/fast/dom/Element/scrollWidth-expected.txt

Test the value of scrollWidth on blocks with visible overflow. These results match IE8.

LTR:
-> scrollWidth: 75
margin-left: 9px; -> scrollWidth: 84
margin-right: 9px; -> scrollWidth: 75
margin-left: -27px; -> scrollWidth: 50
position: relative; -> scrollWidth: 75
position: relative; left: 9px; -> scrollWidth: 84
position: relative; left: -9px; -> scrollWidth: 66
position: relative; left: -27px; -> scrollWidth: 50
position: absolute; -> scrollWidth: 50

RTL:
-> scrollWidth: 75
margin-left: 9px; -> scrollWidth: 75
margin-right: 9px; -> scrollWidth: 84
margin-left: -27px; -> scrollWidth: 75
position: relative; -> scrollWidth: 75
position: relative; left: 9px; -> scrollWidth: 66
position: relative; left: -9px; -> scrollWidth: 84
position: relative; left: -27px; -> scrollWidth: 102
position: absolute; -> scrollWidth: 50
